+++ docs/pages/banner/minimal/+Page.tsx
@@ -1,0 +1,8 @@
+export { Page }
+
+import React from 'react'
+import { BannerMinimal } from '../Banner'
+
+function Page() {
+  return <BannerMinimal />
+}