+++ packages/vike/utils/unique.ts
@@ -1,5 +1,3 @@
-export { unique }
-
-function unique<T>(arr: T[]): T[] {
+export function unique<T>(arr: T[]): T[] {
   return Array.from(new Set(arr))
 }